chore: Extract ClusterInstanceSignIn bulk of logic out too
